requirements.txt
streamlit~=1.36.0
python-dotenv~=1.0.0
pandas~=2.2.0
beautifulsoup4~=4.12.2
